Change operating costs to float

+++ boaviztapi/model/crud_models/configuration_model.py
@@ -23,7 +23,7 @@ class OnPremiseServerUsage(BaseModel):
     avgConsumption: Optional[float] = Field(default=None, ge=1, le=100000)
     serverLoad: Optional[float] = Field(default=None, ge=1, le=100)
     serverLoadAdvanced: Optional[ServerLoadAdvanced] = Field(default=None)
-    operatingCosts: Optional[int] = Field(default=None, ge=0, le=100000000)
+    operatingCosts: Optional[float] = Field(default=None, ge=0, le=100000000)
 
 class OnPremiseConfigurationModel(BaseCRUDModel):
     type: Literal['on-premise']
